fix GetTasks, move common code

+++ common/persistence/sql/task_v1.go
@@ -4,13 +4,10 @@ import (
 	"bytes"
 	"context"
 	"database/sql"
-	"encoding/binary"
 	"fmt"
 	"math"
 
-	"github.com/dgryski/go-farm"
 	commonpb "go.temporal.io/api/common/v1"
-	enumspb "go.temporal.io/api/enums/v1"
 	"go.temporal.io/api/serviceerror"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/persistence"
@@ -47,7 +44,7 @@ func (m *sqlTaskManagerV1) CreateTaskQueue(
 	if err != nil {
 		return serviceerror.NewInternal(err.Error())
 	}
-	tqId, tqHash := m.taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
+	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
 
 	row := sqlplugin.TaskQueuesRow{
 		RangeHash:    tqHash,
@@ -74,7 +71,7 @@ func (m *sqlTaskManagerV1) GetTaskQueue(
 	if err != nil {
 		return nil, serviceerror.NewInternal(err.Error())
 	}
-	tqId, tqHash := m.taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
+	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
 	rows, err := m.Db.SelectFromTaskQueues(ctx, sqlplugin.TaskQueuesFilter{
 		RangeHash:   tqHash,
 		TaskQueueID: tqId,
@@ -112,7 +109,7 @@ func (m *sqlTaskManagerV1) UpdateTaskQueue(
 		return nil, serviceerror.NewInternal(err.Error())
 	}
 
-	tqId, tqHash := m.taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
+	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
 	var resp *persistence.UpdateTaskQueueResponse
 	err = m.SqlStore.txExecute(ctx, "UpdateTaskQueue", func(tx sqlplugin.Tx) error {
 		if err := lockTaskQueue(ctx,
@@ -256,7 +253,7 @@ func (m *sqlTaskManagerV1) DeleteTaskQueue(
 	if err != nil {
 		return serviceerror.NewUnavailable(err.Error())
 	}
-	tqId, tqHash := m.taskQueueIdAndHash(nidBytes, request.TaskQueue.TaskQueueName, request.TaskQueue.TaskQueueType, persistence.SubqueueZero)
+	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue.TaskQueueName, request.TaskQueue.TaskQueueType, persistence.SubqueueZero)
 	result, err := m.Db.DeleteFromTaskQueues(ctx, sqlplugin.TaskQueuesFilter{
 		RangeHash:   tqHash,
 		TaskQueueID: tqId,
@@ -295,7 +292,7 @@ func (m *sqlTaskManagerV1) CreateTasks(
 		if pair, ok := cache[subqueue]; ok {
 			return pair.id, pair.hash
 		}
-		id, hash := m.taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, subqueue)
+		id, hash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, subqueue)
 		cache[subqueue] = pair{id: id, hash: hash}
 		return id, hash
 	}
@@ -336,6 +333,10 @@ func (m *sqlTaskManagerV1) GetTasks(
 	ctx context.Context,
 	request *persistence.GetTasksRequest,
 ) (*persistence.InternalGetTasksResponse, error) {
+	if request.InclusiveMinPass != 0 {
+		return nil, serviceerror.NewInternal("invalid GetTasks request on queue: InclusiveMinPass is not supported")
+	}
+
 	nidBytes, err := primitives.ParseUUID(request.NamespaceID)
 	if err != nil {
 		return nil, serviceerror.NewUnavailable(err.Error())
@@ -351,7 +352,7 @@ func (m *sqlTaskManagerV1) GetTasks(
 		inclusiveMinTaskID = token.TaskID
 	}
 
-	tqId, tqHash := m.taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, request.Subqueue)
+	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, request.Subqueue)
 	rows, err := m.Db.SelectFromTasks(ctx, sqlplugin.TasksFilter{
 		RangeHash:          tqHash,
 		TaskQueueID:        tqId,
@@ -389,11 +390,15 @@ func (m *sqlTaskManagerV1) CompleteTasksLessThan(
 	ctx context.Context,
 	request *persistence.CompleteTasksLessThanRequest,
 ) (int, error) {
+	if request.ExclusiveMaxPass != 0 {
+		return 0, serviceerror.NewInternal("invalid CompleteTasksLessThan request on queue")
+	}
+
 	nidBytes, err := primitives.ParseUUID(request.NamespaceID)
 	if err != nil {
 		return 0, serviceerror.NewUnavailable(err.Error())
 	}
-	tqId, tqHash := m.taskQueueIdAndHash(nidBytes, request.TaskQueueName, request.TaskType, request.Subqueue)
+	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueueName, request.TaskType, request.Subqueue)
 	result, err := m.Db.DeleteFromTasks(ctx, sqlplugin.TasksFilter{
 		RangeHash:          tqHash,
 		TaskQueueID:        tqId,
@@ -408,43 +413,6 @@ func (m *sqlTaskManagerV1) CompleteTasksLessThan(
 		return 0, serviceerror.NewUnavailablef("rowsAffected returned error: %v", err)
 	}
 	return int(nRows), nil
-}
-
-// Returns the persistence task queue id and a uint32 hash for a task queue.
-func (m *sqlTaskManagerV1) taskQueueIdAndHash(
-	namespaceID primitives.UUID,
-	taskQueueName string,
-	taskType enumspb.TaskQueueType,
-	subqueue int,
-) ([]byte, uint32) {
-	id := m.taskQueueId(namespaceID, taskQueueName, taskType, subqueue)
-	return id, farm.Fingerprint32(id)
-}
-
-func (m *sqlTaskManagerV1) taskQueueId(
-	namespaceID primitives.UUID,
-	taskQueueName string,
-	taskType enumspb.TaskQueueType,
-	subqueue int,
-) []byte {
-	idBytes := make([]byte, 0, 16+len(taskQueueName)+1+binary.MaxVarintLen16)
-	idBytes = append(idBytes, namespaceID...)
-	idBytes = append(idBytes, []byte(taskQueueName)...)
-
-	// To ensure that different names+types+subqueue ids never collide, we mark types
-	// containing subqueues with an extra high bit, and then append the subqueue id. There are
-	// only a few task queue types (currently 3), so the high bits are free. (If we have more
-	// fields to append, we can use the next lower bit to mark the presence of that one, etc..)
-	const hasSubqueue = 0x80
-
-	if subqueue > 0 {
-		idBytes = append(idBytes, uint8(taskType)|hasSubqueue)
-		idBytes = binary.AppendUvarint(idBytes, uint64(subqueue))
-	} else {
-		idBytes = append(idBytes, uint8(taskType))
-	}
-
-	return idBytes
 }
 
 func lockTaskQueue(
